+++ example/android/app/src/main/kotlin/com/tolotra/account_picker_example/MainActivity.kt
@@ -1,7 +1,8 @@
 package com.tolotra.account_picker_example
 
+import io.flutter.embedding.android.FlutterActivity
 import io.flutter.embedding.android.FlutterFragmentActivity
 
 // How to use registerForActivityResult in Flutter's MainActivity?
 // https://stackoverflow.com/questions/76020007/how-to-use-registerforactivityresult-in-flutters-mainactivity
-class MainActivity: FlutterFragmentActivity()
+class MainActivity: FlutterActivity()