+++ publish.sh
@@ -1,0 +1,2 @@
+#fvm dart pub publish --dry-run
+fvm dart pub publish